Move contents of LHESource.h to LHESource.cc
